requirements.txt
flask
flask-cors
google-generativeai
pyttsx3==2.90
speechrecognition
pypdf2
pytesseract
opencv-python-headless
numpy
